+++ src/main/java/kr/hhplus/be/server/domain/order/controller/OrderApi.java
@@ -8,11 +8,8 @@ import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import kr.hhplus.be.server.domain.order.dto.OrderRequest;
 import kr.hhplus.be.server.domain.order.dto.OrderResponse;
-import kr.hhplus.be.server.domain.order.dto.PopularProductResponse;
 import kr.hhplus.be.server.global.response.ApiResult;
 import org.springframework.http.ResponseEntity;
-
-import java.util.List;
 
 @Tag(name = "주문 관리", description = "주문을 관리합니다.")
 public interface OrderApi {
@@ -32,10 +29,4 @@ public interface OrderApi {
             @Parameter(description = "주문 요청") OrderRequest orderRequest
     );
 
-    @Operation(summary = "인기 상품 조회", description = "상위 5개의 인기 상품을 조회합니다.")
-    @ApiResponses({
-            @ApiResponse(responseCode = "200", description = "인기 상품 조회 성공", useReturnTypeSchema = true),
-            @ApiResponse(responseCode = "500", description = "서버 오류", content = @Content())
-    })
-    ResponseEntity<ApiResult<List<PopularProductResponse>>> findPopularProducts();
 }
