+++ src/main/java/kr/hhplus/be/server/domain/product/controller/ProductController.java
@@ -1,11 +1,6 @@
 package kr.hhplus.be.server.domain.product.controller;
 
-import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.Parameter;
-import io.swagger.v3.oas.annotations.media.Content;
-import io.swagger.v3.oas.annotations.responses.ApiResponse;
-import io.swagger.v3.oas.annotations.responses.ApiResponses;
-import io.swagger.v3.oas.annotations.tags.Tag;
+import kr.hhplus.be.server.domain.order.dto.PopularProductResponse;
 import kr.hhplus.be.server.domain.product.dto.ProductResponse;
 import kr.hhplus.be.server.global.response.ApiResult;
 import kr.hhplus.be.server.global.response.SuccessCode;
@@ -17,6 +12,8 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+
+import java.util.List;
 
 @RequiredArgsConstructor
 @RestController
@@ -34,6 +31,15 @@ public class ProductController implements ProductApi{
         return ResponseEntity
                 .status(HttpStatus.OK)
                 .body(ApiResult.of(SuccessCode.FIND_PRODUCT, response));
+    }
+
+    @Override
+    @GetMapping("/popular")
+    public ResponseEntity<ApiResult<List<PopularProductResponse>>> findPopularProducts() {
+        List<PopularProductResponse> response = fakeStore.popularProducts();
+        return ResponseEntity
+                .status(HttpStatus.OK)
+                .body(ApiResult.of(SuccessCode.FIND_POPULAR_PRODUCT, response));
     }
 
 }
